restructure validators and dtos
