+++ Halcyon.Api/Features/Account/ForgotPassword/ForgotPasswordDtos.cs
@@ -1,0 +1,6 @@
+namespace Halcyon.Api.Features.Account.ForgotPassword;
+
+public class ForgotPasswordRequest
+{
+    public string EmailAddress { get; set; } = null!;
+}
