+++ Halcyon.Api/Features/Account/ForgotPassword/ForgotPasswordValidators.cs
@@ -2,11 +2,6 @@ using FluentValidation;
 
 namespace Halcyon.Api.Features.Account.ForgotPassword;
 
-public class ForgotPasswordRequest
-{
-    public string EmailAddress { get; set; } = null!;
-}
-
 public class ForgotPasswordRequestValidator : AbstractValidator<ForgotPasswordRequest>
 {
     public ForgotPasswordRequestValidator()
