--- Halcyon.Api/Features/Account/Login/LoginResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace Halcyon.Api.Features.Account.Login;
-
-public class LoginResponse
-{
-    public string AccessToken { get; set; } = null!;
-}
