+++ Halcyon.Api/Features/Account/Login/LoginValidators.cs
@@ -2,13 +2,6 @@
 
 namespace Halcyon.Api.Features.Account.Login;
 
-public class LoginRequest
-{
-    public string EmailAddress { get; set; } = null!;
-
-    public string Password { get; set; } = null!;
-}
-
 public class LoginRequestValidator : AbstractValidator<LoginRequest>
 {
     public LoginRequestValidator()
