+++ Halcyon.Api/Features/Account/Register/RegisterValidators.cs
@@ -3,19 +3,6 @@ using Halcyon.Api.Common.Validation;
 
 namespace Halcyon.Api.Features.Account.Register;
 
-public class RegisterRequest
-{
-    public string EmailAddress { get; set; } = null!;
-
-    public string Password { get; set; } = null!;
-
-    public string FirstName { get; set; } = null!;
-
-    public string LastName { get; set; } = null!;
-
-    public DateOnly DateOfBirth { get; set; }
-}
-
 public class RegisterRequestValidator : AbstractValidator<RegisterRequest>
 {
     public RegisterRequestValidator(TimeProvider timeProvider)
