+++ Halcyon.Api/Features/Account/ResetPassword/ResetPasswordValidators.cs
@@ -2,15 +2,6 @@
 
 namespace Halcyon.Api.Features.Account.ResetPassword;
 
-public class ResetPasswordRequest
-{
-    public Guid Token { get; set; }
-
-    public string EmailAddress { get; set; } = null!;
-
-    public string NewPassword { get; set; } = null!;
-}
-
 public class ResetPasswordRequestValidator : AbstractValidator<ResetPasswordRequest>
 {
     public ResetPasswordRequestValidator()
