+++ Halcyon.Api/Features/CommonDtos.cs
@@ -4,3 +4,8 @@ public class UpdateRequest
 {
     public uint? Version { get; set; }
 }
+
+public class UpdateResponse
+{
+    public Guid Id { get; set; }
+}
