+++ Halcyon.Api/Features/Profile/ChangePassword/ChangePasswordValidators.cs
@@ -2,13 +2,6 @@ using FluentValidation;
 
 namespace Halcyon.Api.Features.Profile.ChangePassword;
 
-public class ChangePasswordRequest : UpdateRequest
-{
-    public string CurrentPassword { get; set; } = null!;
-
-    public string NewPassword { get; set; } = null!;
-}
-
 public class ChangePasswordRequestValidator : AbstractValidator<ChangePasswordRequest>
 {
     public ChangePasswordRequestValidator()
