+++ Halcyon.Api/Features/Profile/UpdateProfile/UpdateProfileValidators.cs
@@ -3,17 +3,6 @@ using Halcyon.Api.Common.Validation;
 
 namespace Halcyon.Api.Features.Profile.UpdateProfile;
 
-public class UpdateProfileRequest : UpdateRequest
-{
-    public string EmailAddress { get; set; } = null!;
-
-    public string FirstName { get; set; } = null!;
-
-    public string LastName { get; set; } = null!;
-
-    public DateOnly DateOfBirth { get; set; }
-}
-
 public class UpdateProfileRequestValidator : AbstractValidator<UpdateProfileRequest>
 {
     public UpdateProfileRequestValidator(TimeProvider timeProvider)
