--- Halcyon.Api/Features/UpdateResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace Halcyon.Api.Features;
-
-public class UpdateResponse
-{
-    public Guid Id { get; set; }
-}
