+++ Halcyon.Api/Features/Users/UpdateUser/UpdateUserDtos.cs
@@ -1,0 +1,14 @@
+namespace Halcyon.Api.Features.Users.UpdateUser;
+
+public class UpdateUserRequest : UpdateRequest
+{
+    public string EmailAddress { get; set; } = null!;
+
+    public string FirstName { get; set; } = null!;
+
+    public string LastName { get; set; } = null!;
+
+    public DateOnly DateOfBirth { get; set; }
+
+    public List<string> Roles { get; set; } = null!;
+}
