+++ Halcyon.Api/Features/Users/UpdateUser/UpdateUserValidators.cs
@@ -3,19 +3,6 @@ using Halcyon.Api.Common.Validation;
 
 namespace Halcyon.Api.Features.Users.UpdateUser;
 
-public class UpdateUserRequest : UpdateRequest
-{
-    public string EmailAddress { get; set; } = null!;
-
-    public string FirstName { get; set; } = null!;
-
-    public string LastName { get; set; } = null!;
-
-    public DateOnly DateOfBirth { get; set; }
-
-    public List<string> Roles { get; set; } = null!;
-}
-
 public class UpdateUserRequestValidator : AbstractValidator<UpdateUserRequest>
 {
     public UpdateUserRequestValidator(TimeProvider timeProvider)
